+++ examples/blink-blocking.rs
@@ -1,7 +1,10 @@
 //! Slowly blink an LED while blocking on a timer.
 //!
 //! Use this as the minimum-viable runtime support. You don't
-//! need interrupts for this example.
+//! need MCU-specific interrupts for this example.
+//!
+//! This example demonstrates how to register an exception
+//! handler. See the API documentation for more information.
 
 #![no_std]
 #![no_main]
@@ -14,5 +17,24 @@ fn main() -> ! {
     loop {
         led.toggle();
         pit.blocking_delay();
+    }
+}
+
+use imxrt_rt::exception;
+
+#[exception]
+unsafe fn DefaultHandler(_irqn: i16) {
+    uh_oh()
+}
+
+#[exception]
+unsafe fn HardFault(_: &imxrt_rt::ExceptionFrame) -> ! {
+    uh_oh()
+}
+
+#[inline(never)]
+fn uh_oh() -> ! {
+    loop {
+        core::sync::atomic::fence(core::sync::atomic::Ordering::SeqCst)
     }
 }