+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Add new MCU targets:
 - imxrt1040
 - imxrt1160
 
+Document workaround for registering exception handlers.
+
 ## [0.1.5] 2024-10-26
 
 Add initial support for RT1180.