+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.1.6] 2025-03-01
+
 Add new MCU targets:
 
 - imxrt1040
@@ -48,7 +50,8 @@ ensure that your build uses this imxrt-rt release.
 First release. `imxrt-rt` provides a build-time API that defines a memory map,
 as well as a runtime library that configures i.MX RT 10xx and 11xx processors.
 
-[Unreleased]: https://github.com/imxrt-rs/imxrt-rt/compare/v0.1.5...HEAD
+[Unreleased]: https://github.com/imxrt-rs/imxrt-rt/compare/v0.1.6...HEAD
+[0.1.6]: https://github.com/imxrt-rs/imxrt-rt/releases/compare/v0.1.5...v0.1.6
 [0.1.5]: https://github.com/imxrt-rs/imxrt-rt/releases/compare/v0.1.4...v0.1.5
 [0.1.4]: https://github.com/imxrt-rs/imxrt-rt/releases/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/imxrt-rs/imxrt-rt/releases/compare/v0.1.2...v0.1.3